backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
geoalchemy2
pydantic
geopandas
shapely
skyfield
python-dateutil
python-multipart